% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQRX.R
\name{create.package.model}
\alias{create.package.model}
\title{Construct an SPQRX model object}
\usage{
create.package.model(
  model,
  n.knots,
  knots,
  hyperparameter = NULL,
  normalizer = NULL,
  variable_names = NULL,
  spqrx = TRUE
)
}
\arguments{
\item{model}{A trained keras model object.}

\item{n.knots}{Integer. Number of spline knots.}

\item{knots}{Numeric vector of knot locations.}

\item{hyperparameter}{Optional list containing model
hyperparameters (e.g., p_a, p_b, c1, c2).}

\item{normalizer}{A list containing normalization parameters.}

\item{variable_names}{Optional character vector of covariate names.}

\item{spqrx}{Logical. TRUE if heavy-tail SPQRX model.}
}
\value{
An object of class `"spqrx_model"`.
}
\description{
Internal constructor used to package a trained SPQR/SPQRX
model together with metadata required for prediction and
explainability methods.
}
\details{
This function should not be called directly by users.
}
\keyword{internal}
