% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQRX.R
\name{create.package.normalize.list}
\alias{create.package.normalize.list}
\title{Construct normalization metadata}
\usage{
create.package.normalize.list(x_std, x_mean, y_max, y_min)
}
\arguments{
\item{x_std}{Numeric vector of feature standard deviations.}

\item{x_mean}{Numeric vector of feature means.}

\item{y_max}{Maximum value of the response variable.}

\item{y_min}{Minimum value of the response variable.}
}
\value{
A list containing normalization parameters.
}
\description{
Internal helper used to store normalization parameters
computed during model fitting.
}
\details{
This function packages the scaling statistics required
to normalize inputs and rescale outputs during prediction.
}
\keyword{internal}
