% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQRX.R
\name{eval.plot.shap}
\alias{eval.plot.shap}
\title{Plot SHAP Values}
\usage{
eval.plot.shap(
  model,
  x_training,
  x_explain,
  y_training,
  y_explain,
  tau = 0.5,
  shapley_method = "empirical"
)
}
\arguments{
\item{model}{A fitted model object returned by \code{fit_spqrx()}.}

\item{x_training}{Matrix or data frame used to train the model.}

\item{x_explain}{Matrix or data frame of covariates to explain.}

\item{y_training}{Vector of responses corresponding to \code{x_training}.}

\item{y_explain}{Vector of responses corresponding to \code{x_explain}.}

\item{tau}{Numeric or vector of quantiles (default 0.5) for QF predictions.}

\item{shapley_method}{Character, the SHAP approximation method (default 'empirical').}
}
\value{
Invisibly returns \code{NULL}. Produces a ggplot2 plot of SHAP feature contributions.
}
\description{
Generates a summary plot of SHAP values for a given model
across features using results from \code{eval.explain.shapr()}.
}
