% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate.simulation.extremes}
\alias{generate.simulation.extremes}
\title{Generate data from Simulation Study in SPQRx}
\usage{
generate.simulation.extremes(n = 1000, n.knots = 25)
}
\arguments{
\item{n}{The data points}
}
\value{
Returns a list
}
\description{
Generate covaraites from uniform distribution
}
