% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQRX.R
\name{xi_custom_activation}
\alias{xi_custom_activation}
\title{Custom activation function for SPQRX xi parameter}
\usage{
xi_custom_activation(x)
}
\arguments{
\item{x}{A tensor input from the previous neural network layer.}
}
\value{
A tensor with values constrained to (0, 0.5).
}
\description{
Internal activation function used in the SPQRX neural network
architecture to constrain the xi parameter to the interval (0, 0.5).
}
\details{
This function applies a sigmoid transformation and rescales
the output by 0.5. For the zeta parameter.
}
\keyword{internal}
